pyproject.toml
[tool.poetry]
name = "codeleap-backend-drf"
version = "0.0.1"
description = "Django-DRF project for Codeleap Engineering team."
authors = ["Eduardo Rezaghi <eduardo.rezaghi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg-binary = "^3.2.3" # use most recent psycopg3
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
django-stubs = "^5.1.1" # Type stubs for Django
djangorestframework-stubs = "^3.15.1" # Type stubs for DRF
psycopg = "^3.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src"
testpaths = ["tests"]